DCAT / Add conversion for the simple URL harvester harvesting DCAT.
This is used to produce DCAT records when using the DCAT harvester.
Records are harvested in DCAT, converted using SPARQL and then user
choose a conversion for the record in the catalogue (ISO or DCAT).

Related to #165316

Some feeds (eg. provided by ODS) were triggering indexing error due to
`skos:Concept` without labels.

eg.
```xml
<dcat:keyword xml:lang="nl">Ivago</dcat:keyword>
<dcat:theme>
  <skos:Concept xmlns:skos="http://www.w3.org/2004/02/skos/core#"
         rdf:about="https://data.stad.gent/id/theme/Natuur%20%26%20milieu">
    <skos:inScheme rdf:resource="https://data.stad.gent/concept-scheme/themes"/>
  </skos:Concept>
</dcat:theme>
```

Some feeds are also using HTML (eg. in abstract), do we need to clean
this?

Do we need to support that?

We can probably remove files in `import/rdf-*` which are now unused.