CSW output configuration for DCAT
`own` outputSchema return the XML as it is in the database. It is not
defined in the CSW spec and is a GeoNetwork specific `outputSchema`.

When a CSW request is made in one of the DCAT related output schema
(added in https://github.com/geonetwork/core-geonetwork/pull/7600)
return the DCAT record as it is.

There is no need to define brief, full and summary conversion when they
are all the same.