Editor / Thesaurus / Add possibility to override configuration.
When a field is using a thesaurus, it has to be configured in the
`config-editor.xml` file in the `editor/fields/for` section

eg.
```xml
<editor>...
  <fields>...
    <for name="dcat:theme" use="thesaurus-list-picker">
      <directiveAttributes
        thesaurus="external.theme.datatheme"
        xpath="/dcat:theme"
        max=""
        labelKey="dcat.addThemes"/>
    </for>
```

User may want to configure thesaurus to be used on a per view basis (eg.
when a view is specific to a profile). In such case, the field
configuration can be overriden a field element.

eg.
```xml
<field xpath="/rdf:RDF/dcat:Catalog/dcat:dataset/dcat:Dataset/dcat:theme"
       name="dcat:theme"
       or="theme"
       use="data-gn-keyword-picker"
       in="/rdf:RDF/dcat:Catalog/dcat:dataset/dcat:Dataset">
  <directiveAttributes
    thesaurus="external.theme.vl-datatheme"
    xpath="/dcat:theme"
    max=""
    labelKey="dcat-addThemes"/>
</field>
```